+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+test
 # Unreleased
 
 # 3.2.0 / 30-10-2025