+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [IMPROVEMENT] Add missing `versionMajor` property to the `DDLogEventOperatingSystem` definition in Objective-C. See [#2463][]
 - [IMPROVEMENT] Add `ddtags` to RUM events. See [#2436][]
 - [FIX] Fix `LogEvent` device types. See [#2474][]
+- [BUGFIX] Fix tracing header injection for sampled out requests. See [#2473][]
 
 # 3.0.0 / 02-09-2025
 
@@ -979,6 +980,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2538]: https://github.com/DataDog/dd-sdk-ios/pull/2538
 [#2532]: https://github.com/DataDog/dd-sdk-ios/pull/2532
 [#2514]: https://github.com/DataDog/dd-sdk-ios/pull/2514
+[#2473]: https://github.com/DataDog/dd-sdk-ios/pull/2473
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin