+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
-- [BUGFIX] Fix tracing header injection for sampled out requests. See [#2473][]
 
+# 3.3.0 / 17-11-2025
+
+- [FIX] Fix tracing header injection for sampled out requests. See [#2473][]
+- [FIX] Remove GraphQL headers from request after processing it. See [#2566][]
 - [FEATURE] Support errors for GraphQL requests. See [#2552]
 
 # 3.2.0 / 30-10-2025
@@ -984,6 +987,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2532]: https://github.com/DataDog/dd-sdk-ios/pull/2532
 [#2514]: https://github.com/DataDog/dd-sdk-ios/pull/2514
 [#2552]: https://github.com/DataDog/dd-sdk-ios/pull/2552
+[#2566]: https://github.com/DataDog/dd-sdk-ios/pull/2566
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin